Task 6/requirements.txt
discord.py>=2.3.2
aiohttp>=3.8.0
python-dotenv>=1.0.0